Move modular-edge + cycle-removal out of GraphBuilder into modular_pipelines
SRP cleanup from the bb500756 review: add_modular_edges, remove_cyclic_modular_
edges and _reachable_from build/clean the modular tree's edges, which is the
modular_pipelines module's concern, not the graph builder's. They were already
pure static functions, so the move is mechanical.

- modular_pipelines.py: add the three functions (next to ModularTreeBuilder)
  plus the imports they need (defaultdict, GraphEdgeAPIResponse). Also adopt
  node_ids.task_node_id_for here.
- graph_builder.py: import and call the relocated functions; drop the two static
  methods and the module-level _reachable_from; drop the now-unused
  ROOT_MODULAR_PIPELINE_ID import.
- test_graph_builder.py: call the relocated remove_cyclic_modular_edges.

Gate: make lint clean; pytest package/tests/ 563 passed.
