+++ package/kedro_viz/integrations/kedro/inspection/graph_builder.py
@@ -29,15 +29,14 @@ from kedro_viz.api.rest.responses.pipelines import (
     NamedEntityAPIResponse,
     TaskNodeAPIResponse,
 )
-from kedro_viz.constants import (
-    DEFAULT_REGISTERED_PIPELINE_ID,
-    ROOT_MODULAR_PIPELINE_ID,
-)
+from kedro_viz.constants import DEFAULT_REGISTERED_PIPELINE_ID
 from kedro_viz.integrations.kedro import node_ids
 from kedro_viz.integrations.kedro.inspection.modular_pipelines import (
     ModularMembership,
     ModularTreeBuilder,
     ModularTreeEntry,
+    add_modular_edges,
+    remove_cyclic_modular_edges,
 )
 from kedro_viz.models.flowchart.model_utils import GraphNodeType
 from kedro_viz.services.layers import sort_layers
@@ -150,8 +149,8 @@ class GraphBuilder:
         tree_builder = ModularTreeBuilder(pipeline.nodes)
         tree = tree_builder.build()
         nodes.extend(self._build_modular_pipeline_nodes(tree_builder, selected))
-        self._add_modular_edges(edges, tree)
-        self._remove_cyclic_modular_edges(edges, tree)
+        add_modular_edges(edges, tree)
+        remove_cyclic_modular_edges(edges, tree)
 
         return GraphAPIResponse(
             nodes=nodes,
@@ -250,43 +249,6 @@ class GraphBuilder:
         return sort_layers(cast("dict[str, GraphNode]", nodes_by_id), dependencies)
 
     @staticmethod
-    def _add_modular_edges(
-        edges: dict[tuple[str, str], GraphEdgeAPIResponse],
-        tree: dict[str, ModularTreeEntry],
-    ) -> None:
-        """Connect each modular pipeline to its boundary datasets (input -> mp, mp -> output)."""
-        for mp_id, entry in tree.items():
-            if mp_id == ROOT_MODULAR_PIPELINE_ID:
-                continue
-            for input_id in entry.inputs:
-                edges.setdefault(
-                    (input_id, mp_id),
-                    GraphEdgeAPIResponse(source=input_id, target=mp_id),
-                )
-            for output_id in entry.outputs:
-                edges.setdefault(
-                    (mp_id, output_id),
-                    GraphEdgeAPIResponse(source=mp_id, target=output_id),
-                )
-
-    @staticmethod
-    def _remove_cyclic_modular_edges(
-        edges: dict[tuple[str, str], GraphEdgeAPIResponse],
-        tree: dict[str, ModularTreeEntry],
-    ) -> None:
-        """Drop any ``input -> mp`` edge whose input is also reachable *from* the mp (a cycle)."""
-        adjacency: dict[str, set[str]] = defaultdict(set)
-        for source, target in edges:
-            adjacency[source].add(target)
-        for mp_id, entry in tree.items():
-            if mp_id == ROOT_MODULAR_PIPELINE_ID:
-                continue
-            reachable = _reachable_from(mp_id, adjacency)
-            for input_id in entry.inputs & reachable:
-                edges.pop((input_id, mp_id), None)
-                adjacency[input_id].discard(mp_id)
-
-    @staticmethod
     def _register_dataset(
         name: str,
         node: NodeSnapshot,
@@ -304,19 +266,6 @@ class GraphBuilder:
         edges.setdefault(
             (source, target), GraphEdgeAPIResponse(source=source, target=target)
         )
-
-
-def _reachable_from(start: str, adjacency: dict[str, set[str]]) -> set[str]:
-    """Return all nodes reachable from ``start`` (excluding ``start`` unless it is in a cycle)."""
-    seen: set[str] = set()
-    stack = list(adjacency.get(start, ()))
-    while stack:
-        node = stack.pop()
-        if node in seen:
-            continue
-        seen.add(node)
-        stack.extend(adjacency.get(node, ()))
-    return seen
 
 
 def _to_tree_response(
